Crawl_API/requirements.txt
fastapi~=0.121.2
mysql-connector-python~=9.5.0
pymongo~=4.15.4
python-dotenv~=1.2.1
requests~=2.32.5
regex~=2025.9.18
beautifulsoup4~=4.14.2
uvicorn~=0.38.0
pip~=25.2
typing_extensions~=4.15.0
attrs~=25.4.0
distro~=1.9.0
protobuf~=5.29.5
filelock~=3.13.1
PyJWT~=2.10.1
starlette~=0.49.3